transition to knex
