+++ package.json
@@ -4,16 +4,14 @@
     "description": "An interstate delivcery platform for people to send items to other geographical states through individuals travelling to those locations.",
     "main": "app.js",
     "scripts": {
-        "lint": "eslint . --ext .ts --fix",
+        "start": "node dist/app.js",
+        "lint": "eslint --ext .ts .--fix",
         "prettier": "prettier --config .prettierrc --ignore-path .gitignore  --write './src/**/*.+(js|ts)'",
-        "format": "npm run lint && npm run prettier ",
-        "dev": "npm run tsc && nodemon ./src/app.ts",
-        "jasmine": "jasmine",
-        "test": "NODE_ENV=test db-migrate -e test up && NODE_ENV=test npm run jasmine",
-        "test-mgt": "NODE_ENV=test db-migrate --env test up",
-        "create-test-db": "NODE_ENV=test db-migrate db:create deliverycog_test",
-        "drop-test-db": "NODE_ENV=test db-migrate db:drop deliverycog_test",
-        "tsc": "tsc"
+        "format": "npm run prettier ",
+        "migrate": "npx knex migrate:latest --knexfile ./src/V1/APIs/Config/knexfile.ts",
+        "down": "npx knex migrate:down  --knexfile ./src/V1/APIs/Config/knexfile.ts",
+        "dev": "npm run build && nodemon ./src/app.ts",
+        "build": "tsc"
     },
     "repository": {
         "type": "git",
@@ -55,13 +53,14 @@
     },
     "dependencies": {
         "bcrypt": "^5.0.1",
-        "db-migrate-pg": "^1.2.2",
         "dotenv": "^16.0.3",
         "express": "^4.18.2",
         "express-validator": "^6.14.2",
         "html-to-text": "^8.2.1",
         "jsonwebtoken": "^8.5.1",
+        "knex": "^2.3.0",
         "nodemailer": "^6.7.8",
+        "objection": "^3.0.1",
         "pg": "^8.8.0",
         "swagger-jsdoc": "^6.2.5",
         "swagger-ui-express": "^4.5.0"
