+++ src/V1/APIs/Config/db/dbSetup.ts
@@ -1,0 +1,11 @@
+import dotenv from 'dotenv';
+import knex from 'knex';
+import knexfile from '../knexfile';
+import { Model } from 'objection';
+
+const setupDb = () => {
+    const db = knex(knexfile.development);
+    Model.knex(db);
+};
+
+export default setupDb;
