+++ src/V1/APIs/Docs/auth.docs.ts
