+++ src/V1/APIs/Models/User.ts
@@ -1,0 +1,19 @@
+import { Model } from 'objection';
+
+export interface UserType {
+    id?: number;
+    first_name: string;
+    last_name: string;
+    password_digest: string;
+    phone_number: number;
+    email: string;
+    verification_code?: number;
+    is_verified?: boolean;
+    created_at?: string;
+}
+
+export class User extends Model {
+    static get tableName() {
+        return 'users';
+    }
+}
