+++ src/V1/APIs/Repository/auth/auth.repository.ts
@@ -1,0 +1,34 @@
+import { Encryption } from '../../Utilities/bcrypt';
+import { User, UserType } from '../../Models/User';
+
+export default class AuthRepository {
+    async createUser(user: UserType): Promise<UserType> {
+        const hashPassword = await new Encryption().bcrypt(
+            user.password_digest
+        );
+        const newUser: any = await User.query().insert({
+            first_name: user.first_name,
+            last_name: user.last_name,
+            password_digest: hashPassword,
+            phone_number: user.phone_number,
+            email: user.email,
+            verification_code: user.verification_code,
+        });
+        return newUser;
+    }
+    async authenticate(
+        email: string,
+        password: string
+    ): Promise<UserType[] | undefined> {
+        const user: any = await User.query().where('email', '=', email);
+        const checkPassword = await new Encryption().compare(
+            password,
+            user[0].password_digest
+        );
+
+        if (!checkPassword) {
+            return undefined;
+        }
+        return user;
+    }
+}
