+++ src/V1/APIs/Routes/index.ts
@@ -1,0 +1,12 @@
+import { Router } from 'express';
+import authRouter from './auth/auth.router';
+import swaggerDoc from 'swagger-ui-express';
+import swaggerDocumentation from '../Docs/swagger.config';
+
+const router = Router();
+
+router.use('/auth', authRouter);
+router.use('/documentation', swaggerDoc.serve);
+router.use('/documentation', swaggerDoc.setup(swaggerDocumentation));
+
+export default router;
