--- src/v1/apis/Middlewares/validateRequest.middleware.ts
@@ -1,19 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-import { validationResult } from 'express-validator';
-
-export const validate = (req: Request, res: Response, next: NextFunction) => {
-    const errors = validationResult(req);
-    if (errors.isEmpty()) {
-        return next();
-    }
-
-    const resultErrors = [];
-    errors.array().map((err) => resultErrors.push({ [err.param]: err.msg }));
-
-    resultErrors.push({ message: 'Action unsuccessful' });
-    resultErrors.push({ success: false });
-
-    const errorObject = Object.assign({}, ...resultErrors);
-    return res.status(422).json(errorObject);
-};
-
