--- src/v1/apis/Models/globalQueries.ts
@@ -1,7 +1,0 @@
-export interface Query {
-    model: string;
-    table: string;
-    value?: string | number;
-    uniqueColumn?: string;
-    uniqueValue?: string | number;
-}
